feat(#186): display version on splash screen
