--- apps/electron/src/assets/gitkeep_tmpl_
@@ -1,1 +1,0 @@
-Assets Folder
